project refactoring for clear structure
