+++ src/controller/command/create/strategy/AllDayEventCreator.java
@@ -1,44 +1,42 @@
-package controller.command.create;
+package controller.command.create.strategy;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 
 import model.event.Event;
 import model.exceptions.InvalidEventException;
 import utilities.DateTimeUtil;
 
 /**
- * Strategy for creating a single event.
+ * Strategy for creating an all-day event.
  * Extends AbstractEventCreator to inherit common functionality.
  */
-public class SingleEventCreator extends AbstractEventCreator {
+public class AllDayEventCreator extends AbstractEventCreator {
 
   private final String eventName;
-  private final LocalDateTime startDateTime;
-  private final LocalDateTime endDateTime;
+  private final LocalDate date;
+  private final boolean autoDecline;
   private final String description;
   private final String location;
   private final boolean isPublic;
-  private final boolean autoDecline;
 
   /**
-   * Constructs a strategy for creating a single event.
+   * Constructs a strategy for creating an all-day event.
    *
    * @param args the arguments for event creation
    */
-  public SingleEventCreator(String[] args) {
+  public AllDayEventCreator(String[] args) {
     if (args.length < 4) {
-      throw new IllegalArgumentException("Insufficient arguments for creating a single event");
+      throw new IllegalArgumentException("Insufficient arguments for creating an all-day event");
     }
 
     try {
       this.eventName = args[1];
-      this.startDateTime = DateTimeUtil.parseDateTime(args[2]);
-      this.endDateTime = DateTimeUtil.parseDateTime(args[3]);
+      this.date = DateTimeUtil.parseDate(args[2]);
+      this.autoDecline = Boolean.parseBoolean(args[3]);
 
       this.description = args.length > 4 ? removeQuotes(args[4]) : null;
       this.location = args.length > 5 ? removeQuotes(args[5]) : null;
       this.isPublic = args.length > 6 ? Boolean.parseBoolean(args[6]) : true;
-      this.autoDecline = args.length > 7 ? Boolean.parseBoolean(args[7]) : false;
     } catch (Exception e) {
       throw new IllegalArgumentException("Error parsing arguments: " + e.getMessage(), e);
     }
@@ -48,19 +46,12 @@ public class SingleEventCreator extends AbstractEventCreator {
   public Event createEvent() throws InvalidEventException {
     validateEventParameters(eventName);
 
-    if (startDateTime == null) {
-      throw new InvalidEventException("Start date/time cannot be null");
+    if (date == null) {
+      throw new InvalidEventException("Date cannot be null");
     }
 
-    // Create an event directly
-    return new Event(
-            eventName,
-            startDateTime,
-            endDateTime,
-            description,
-            location,
-            isPublic
-    );
+    return Event.createAllDayEvent(
+            eventName, date, description, location, isPublic);
   }
 
   @Override
@@ -70,6 +61,6 @@ public class SingleEventCreator extends AbstractEventCreator {
 
   @Override
   protected String getSuccessMessage(Event event) {
-    return "Event '" + eventName + "' created successfully.";
+    return "All-day event '" + eventName + "' created successfully.";
   }
 }
