+++ src/controller/command/create/strategy/AllDayRecurringEventCreator.java
@@ -1,4 +1,4 @@
-package controller.command.create;
+package controller.command.create.strategy;
 
 import java.time.LocalDate;
