+++ src/controller/command/create/strategy/AllDayRecurringUntilEventCreator.java
@@ -1,4 +1,4 @@
-package controller.command.create;
+package controller.command.create.strategy;
 
 import java.time.LocalDate;
