+++ src/controller/command/create/strategy/EventCreator.java
@@ -1,4 +1,4 @@
-package controller.command.create;
+package controller.command.create.strategy;
 
 import model.calendar.ICalendar;
 import model.event.Event;
