+++ src/controller/command/edit/strategy/AllEventsEditor.java
@@ -1,4 +1,4 @@
-package controller.command.edit;
+package controller.command.edit.strategy;
 
 import model.calendar.ICalendar;
 import model.exceptions.ConflictingEventException;
