+++ src/controller/command/edit/strategy/SingleEventEditor.java
@@ -1,4 +1,4 @@
-package controller.command.edit;
+package controller.command.edit.strategy;
 
 import java.time.LocalDateTime;
