+++ src/controller/command/event/CommandFactory.java
@@ -1,9 +1,12 @@
-package controller.command;
+package controller.command.event;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import controller.ICommandFactory;
+import controller.command.CommandAdapter;
+import controller.command.CommandExecutor;
+import controller.command.ICommand;
 import controller.command.create.CreateEventCommand;
 import controller.command.edit.EditEventCommand;
 import model.calendar.ICalendar;
